refactor: abstract rich text file upload logic into a reusable handler hook
