+++ apps/web/app/modules/Admin/EditCourse/CourseLessons/NewLesson/ContentLessonForm/ContentLessonForm.tsx
@@ -1,15 +1,6 @@
-import {
-  ALLOWED_EXCEL_FILE_TYPES,
-  ALLOWED_LESSON_IMAGE_FILE_TYPES,
-  ALLOWED_PDF_FILE_TYPES,
-  ALLOWED_PRESENTATION_FILE_TYPES,
-  ALLOWED_VIDEO_FILE_TYPES,
-  ALLOWED_WORD_FILE_TYPES,
-  ENTITY_TYPES,
-} from "@repo/shared";
+import { ENTITY_TYPES } from "@repo/shared";
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { match } from "ts-pattern";
 
 import { useInitializeLessonContext } from "~/api/mutations/admin/useInitializeLessonContext";
 import { useInitVideoUpload } from "~/api/mutations/admin/useInitVideoUpload";
@@ -22,10 +13,10 @@ import { Label } from "~/components/ui/label";
 import { useToast } from "~/components/ui/use-toast";
 import { useLeaveModal } from "~/context/LeaveModalContext";
 import {
-  buildEntityResourceUrl,
-  insertResourceIntoEditor,
-  useEntityResourceUpload,
-} from "~/hooks/useEntityResourceUpload";
+  buildRichTextFileUploadHandler,
+  RICH_TEXT_ACCEPTED_FILE_TYPES,
+} from "~/hooks/buildRichTextFileUploadHandler";
+import { useEntityResourceUpload } from "~/hooks/useEntityResourceUpload";
 import { useTusVideoUpload } from "~/hooks/useTusVideoUpload";
 import { useUploadDisplayModeDialog } from "~/hooks/useUploadDisplayModeDialog";
 import DeleteConfirmationModal from "~/modules/Admin/components/DeleteConfirmationModal";
@@ -39,7 +30,6 @@ import { useContentLessonForm } from "./hooks/useContentLessonForm";
 
 import type { Chapter, Lesson } from "../../../EditCourse.types";
 import type { SupportedLanguages } from "@repo/shared";
-import type { Editor as TiptapEditor } from "@tiptap/react";
 
 type ContentLessonProps = {
   setContentTypeToDisplay: (contentTypeToDisplay: string) => void;
@@ -88,115 +78,40 @@ const ContentLessonForm = ({
     setIsModalOpen(true);
   };
 
-  const removeVideoEmbedBySource = (editor: TiptapEditor, src: string) => {
-    editor
-      .chain()
-      .focus()
-      .command(({ state, tr, dispatch }) => {
-        let found = false;
-
-        state.doc.descendants((node, pos) => {
-          if (node.type.name !== "video") return true;
-          if (node.attrs.src !== src) return true;
-
-          tr.delete(pos, pos + node.nodeSize);
-          found = true;
-          return false;
-        });
-
-        if (!found) return false;
-
-        dispatch?.(tr);
-        return true;
-      })
-      .run();
-  };
-
-  const handleFileUpload = async (file?: File, editor?: TiptapEditor | null) => {
-    if (!file) return;
-
-    const isVideo = ALLOWED_VIDEO_FILE_TYPES.includes(file.type);
-    const isPresentation = ALLOWED_PRESENTATION_FILE_TYPES.includes(file.type);
-    const isPdf = ALLOWED_PDF_FILE_TYPES.includes(file.type);
-    const isDocument =
-      isPdf ||
-      ALLOWED_EXCEL_FILE_TYPES.includes(file.type) ||
-      ALLOWED_WORD_FILE_TYPES.includes(file.type);
-
-    if (isVideo) {
-      let insertedResourceUrl: string | null = null;
-
-      try {
-        const session = await getSessionForFile({
-          file,
-          init: () =>
-            initVideoUpload({
-              filename: file.name,
-              sizeBytes: file.size,
-              mimeType: file.type,
-              title: file.name,
-              resource: "lesson-content",
-              contextId,
-              entityId: lessonToEdit?.id,
-              entityType: ENTITY_TYPES.LESSON,
-            }),
-        });
-
-        if (session.resourceId && editor) {
-          insertedResourceUrl = buildEntityResourceUrl(session.resourceId, ENTITY_TYPES.LESSON);
-
-          editor
-            .chain()
-            .insertContent("<br />")
-            .setVideoEmbed({
-              src: insertedResourceUrl,
-              sourceType: session.provider === "s3" ? "external" : "internal",
-            })
-            .run();
-        }
-
-        await uploadVideo({ file, session });
-      } catch (error) {
-        if (editor && insertedResourceUrl) removeVideoEmbedBySource(editor, insertedResourceUrl);
-
-        console.error("Error uploading video:", error);
-        toast({
-          description: t("uploadFile.toast.videoFailed"),
-          variant: "destructive",
-        });
-      }
-      return;
-    }
-
-    let displayMode: "preview" | "download" = "preview";
-
-    if (isPresentation || isPdf) {
-      const selectedMode = await askForDisplayMode(file.name);
-      if (!selectedMode) return;
-      displayMode = selectedMode;
-    }
-
-    const resourceId = await uploadResource({
-      file,
-      entityType: ENTITY_TYPES.LESSON,
-      entityId: lessonToEdit?.id,
-      contextId,
-      language,
-    });
-
-    insertResourceIntoEditor({
-      editor,
-      resourceId,
-      entityType: ENTITY_TYPES.LESSON,
-      file,
-      resourceType: match({ isPresentation, isPdf, isDocument })
-        .with({ isPresentation: true }, () => "presentation" as const)
-        .with({ isPdf: true }, () => "pdf" as const)
-        .with({ isDocument: true }, () => "document" as const)
-        .otherwise(() => "other" as const),
-      displayMode: isPresentation || isDocument ? displayMode : "preview",
-    });
-  };
+  const handleFileUpload = buildRichTextFileUploadHandler({
+    entityType: ENTITY_TYPES.LESSON,
+    getVideoSessionForFile: (file) =>
+      getSessionForFile({
+        file,
+        init: () =>
+          initVideoUpload({
+            filename: file.name,
+            sizeBytes: file.size,
+            mimeType: file.type,
+            title: file.name,
+            resource: "lesson-content",
+            contextId,
+            entityId: lessonToEdit?.id,
+            entityType: ENTITY_TYPES.LESSON,
+          }),
+      }),
+    uploadVideo,
+    uploadResourceFile: (file) =>
+      uploadResource({
+        file,
+        entityType: ENTITY_TYPES.LESSON,
+        entityId: lessonToEdit?.id,
+        contextId,
+        language,
+      }),
+    askForDisplayMode,
+    onVideoUploadError: () => {
+      toast({
+        description: t("uploadFile.toast.videoFailed"),
+        variant: "destructive",
+      });
+    },
+  });
 
   const missingTranslations =
     lessonToEdit && !lessonToEdit.title.trim() && !lessonToEdit.description.trim();
@@ -293,14 +208,7 @@ const ContentLessonForm = ({
                     content={field.value}
                     lessonId={lessonToEdit?.id}
                     allowFiles={!!lessonToEdit?.id || !!contextId}
-                    acceptedFileTypes={[
-                      ...ALLOWED_LESSON_IMAGE_FILE_TYPES,
-                      ...ALLOWED_VIDEO_FILE_TYPES,
-                      ...ALLOWED_EXCEL_FILE_TYPES,
-                      ...ALLOWED_PDF_FILE_TYPES,
-                      ...ALLOWED_WORD_FILE_TYPES,
-                      ...ALLOWED_PRESENTATION_FILE_TYPES,
-                    ]}
+                    acceptedFileTypes={RICH_TEXT_ACCEPTED_FILE_TYPES}
                     onUpload={handleFileUpload}
                     uploadProgress={isUploading ? (uploadProgress ?? 0) : null}
                     onCtrlSave={() => form.handleSubmit(onSubmit)()}
